This PR relands #110022 but accounts for the changes in #110191. Also, the function for creating COW storages is called `lazy_clone_storage` in this PR, instead of `try_ensure`

NOTE: COW storages do not actually copy on write yet, they just have the COW deleter and deleter context applied to them

Part of #109833



cc ezyang bhosmer smessmer ljk53 bdhirsh

[ghstack-poisoned]

+++ c10/core/impl/cow/COW.cpp
@@ -33,7 +33,7 @@ at::DataPtr copy_data_ptr(at::DataPtr const& data_ptr) {
 
 bool is_simple_data_ptr(const c10::DataPtr& data_ptr) {
 #ifdef C10_MOBILE
-  return data_ptr.get() == data_ptr.get_context() + c10::gAlignment;
+  return static_cast<size_t>(data_ptr.get()) == static_cast<size_t>(data_ptr.get_context()) + c10::gAlignment;
 #else
   return data_ptr.get() == data_ptr.get_context();
 #endif
